refactor: 구독알림 그룹 버튼 컴포넌트 제작
