+++ packages/mobile/src/components/templates/FullPageModalTemplate/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useLocation, useNavigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import $ from "./style.module.scss";
 
@@ -12,9 +12,7 @@ type Props = {
 
 function FullPageModalTemplate({ left, title, right, children }: Props) {
   const navigate = useNavigate();
-  const { pathname } = useLocation();
 
-  // TODO: 분기처리 지우기.. 모달에서 X누르면 그 전 history state를 삭제가 안되서 이렇게 작업
   return (
     <div className={$["full-modal"]}>
       <div className={$.header}>
@@ -22,9 +20,6 @@ function FullPageModalTemplate({ left, title, right, children }: Props) {
           type="button"
           className={$.left}
           onClick={() => {
-            if (pathname === "/setting/subscription") {
-              return navigate("/home");
-            }
             return navigate(-1);
           }}
         >
