+++ packages/mobile/src/main.tsx
@@ -6,6 +6,7 @@ import Toast from "@components/atoms/Toast";
 import { apiServer } from "@shared/constant";
 import { OpenAPI } from "@shared/swagger-api/generated/";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { PersistGate } from "redux-persist/integration/react";
 
 import App from "./App";
@@ -24,6 +25,7 @@ export const queryClient = new QueryClient({
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <Provider store={store}>
     <QueryClientProvider client={queryClient}>
+      <ReactQueryDevtools initialIsOpen />
       <React.StrictMode>
         <PersistGate loading={null} persistor={persistor}>
           <App />
