+++ packages/mobile/src/page/Board/Article/index.tsx
@@ -1,21 +1,10 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
 
-import {
-  Alarm,
-  LeftArrow,
-  Subscription,
-  UnAlarm,
-  UnSubscription,
-} from "@components/atoms/icon";
+import { LeftArrow } from "@components/atoms/icon";
+import SubscriptionNoticeGroup from "@components/shared/SubscriptionNoticeGroup";
 import FullPageModalTemplate from "@components/templates/FullPageModalTemplate";
-import {
-  useBoardQuery,
-  useNoticeBoardMutation,
-  useSubscribeBoardMutation,
-  useUnNoticeBoardMutation,
-  useUnSubscribeBoardMutation,
-} from "@hooks/api/board";
+import { useBoardQuery } from "@hooks/api/board";
 import ArticleList from "src/page/Article/components/ArticleList";
 
 import $ from "./style.module.scss";
@@ -24,30 +13,11 @@ function BoardArticle() {
   const id = Number(useLocation().pathname.split("/").at(-1));
   const uuid = "1111";
   const { data: boardData } = useBoardQuery({ id, uuid });
-  const postSubscribeBoard = useSubscribeBoardMutation({ id, uuid });
-  const deleteSubscribeBoard = useUnSubscribeBoardMutation({ id, uuid });
-  const postNoticeBoard = useNoticeBoardMutation({ id, uuid });
-  const deleteUnNoticeBoard = useUnNoticeBoardMutation({ id, uuid });
 
   if (!boardData) return <></>;
 
-  const handleUnNoticeClick = () => {
-    deleteUnNoticeBoard.mutate({ id, uuid });
-  };
-
-  const handleNoticeClick = () => {
-    postNoticeBoard.mutate({ id, uuid });
-  };
-
-  const handleUnSubscriptionClick = () => {
-    deleteSubscribeBoard.mutate({ id, uuid });
-  };
-
-  const handleSubscriptionClick = () => {
-    postSubscribeBoard.mutate({ id, uuid });
-  };
-
   const { name, isNotice, isSubscribe } = boardData;
+
   return (
     <div className={$["board-article"]}>
       <FullPageModalTemplate
@@ -59,26 +29,11 @@ function BoardArticle() {
         }
         right={
           <div className={$.right}>
-            {isSubscribe ? (
-              <div className={$.icon}>
-                <button type="button" onClick={handleUnSubscriptionClick}>
-                  <Subscription size={30} stroke="#D66D6E" />
-                </button>
-                {isNotice ? (
-                  <button type="button" onClick={handleUnNoticeClick}>
-                    <Alarm size={22} stroke="#D66D6E" />
-                  </button>
-                ) : (
-                  <button type="button" onClick={handleNoticeClick}>
-                    <UnAlarm size={22} stroke="#AAAAAA" />
-                  </button>
-                )}
-              </div>
-            ) : (
-              <button type="button" onClick={handleSubscriptionClick}>
-                <UnSubscription size={30} stroke="#AAAAAA" />
-              </button>
-            )}
+            <SubscriptionNoticeGroup
+              id={id}
+              isNotice={isNotice ?? false}
+              isSubscribe={isSubscribe ?? false}
+            />
           </div>
         }
       >
