+++ packages/mobile/src/page/Board/Article/style.module.scss
@@ -9,19 +9,6 @@
     }
   }
 
-  .right {
-    .icon {
-      display: flex;
-
-      button {
-        margin-right: 8px;
-
-        &:last-child {
-          margin-right: 0;
-        }
-      }
-    }
-  }
 
   .article-list-wrap {
     margin-top: 8px;
