+++ packages/mobile/src/page/Setting/Subscription/style.module.scss
@@ -1,0 +1,15 @@
+.subscription {
+  padding-top: 8px;
+
+  .item {
+    height: 102px;
+    padding: 20px;
+    margin-bottom: 8px;
+    background: $black-0;
+    border-radius: 20px;
+
+    &:last-child {
+      margin-bottom: 0;
+    }
+  }
+}
